refactor(bootstrap): move bootstrap templates to `ui-bootstrap` folder
