--- src/templates.ts
@@ -1,7 +1,0 @@
-export * from './templates/types/types';
-export * from './templates/wrappers/wrappers';
-export { FormlyBootstrapModule } from './templates/formlyBootstrap';
-export { Field } from './templates/field';
-export { FieldType } from './templates/field.type';
-export { FieldWrapper } from './templates/field.wrapper';
-export { SingleFocusDispatcher } from './services/formly.single.focus.dispatcher'
