+++ src/ui-bootstrap/types/checkbox.ts
@@ -1,7 +1,6 @@
 import { Component } from '@angular/core';
 import { FormControl, AbstractControl } from '@angular/forms';
-import { FormlyFieldConfig } from '../../components/formly.field.config';
-import { FieldType } from '../field.type';
+import { FieldType, FormlyFieldConfig } from '../../core/core';
 
 @Component({
   selector: 'formly-field-checkbox',
