+++ src/ui-bootstrap/ui-bootstrap.config.ts
@@ -1,4 +1,4 @@
-import { ConfigOption } from '../services/formly.config';
+import { ConfigOption } from '../core/services/formly.config';
 import {
   FormlyFieldInput,
   FormlyFieldCheckbox,
