+++ src/ui-bootstrap/ui-bootstrap.ts
@@ -1,0 +1,4 @@
+export * from './types/types';
+export * from './wrappers/wrappers';
+export { FormlyValidationMessage } from './formly.validation-message';
+export { FormlyBootstrapModule } from './ui-bootstrap.module';
